+++ NEWS.md
@@ -1,3 +1,7 @@
+## Changes in portfolioBacktest version 0.3.1.9000 (2021-XX-XX)
+
+* Added function backtestChartSharpeRatio()
+
 
 ## Changes in portfolioBacktest version 0.3.1 (2021-10-11)